+++ sum_harness_instructional 656/sum_vector.cpp
@@ -23,7 +23,7 @@ sum(int64_t N, uint64_t A[])
 {
    printf(" inside sum_vector perform_sum, N=%lld \n", N);
 
-    int sum = 0;
+    int64_t sum = 0;
 
     for (int i=0; i < N; i++) {
         sum += A[i];
